move components for pages to pages folders
