+++ components/pages/activity/components/ActivityPagination/styled.js
