+++ components/pages/activity/components/ActivityTable/baseCellStyles.js
