+++ components/pages/login/LoginPage.js
@@ -1,11 +1,13 @@
 import React from 'react';
 
 import WithAuth from 'lib/auth/withAuth';
-import { NotifierProvider } from 'contexts/NotifierContext';
 import { withApolloClient } from 'lib/withApolloClient';
+import { NotifierProvider } from 'contexts/NotifierContext';
+
 import DefaultTemplate from 'components/shared/templates/DefaultTemplate';
-import LoginForm from 'components/shared/organisms/LoginForm';
 import Notifier from 'components/shared/atoms/Notifier';
+
+import LoginForm from './components/LoginForm';
 
 const LoginPage = () => {
   return (
